refactor SimpleDirectoryReader parsers
